+++ R/dynloglikMC.R
@@ -6,15 +6,16 @@
 #' @param y vector: points where the function is evaluated.
 #' @param nreps non-negative integer: number of replications to be used in the computation of the integral in the normalizing
 #' constant.
+#' @param xiInst non-negative real: shape parameter of the instrumental GPD.
+#' @param betaInst non-negative real: scale parameter of the instrumental GPD.
 #' @param weight 'cau' or 'exp': name of weight distribution.
 #' @return Log-likelihood of the lognormal-GPD mixture evaluated at y.
 #' @keywords dynamic mixture.
 #' @export
 #' @examples
-#' llik <- dynloglikMC(c(1,2,0,1,.25,3.5),Metro2019,10000)
+#' llik <- dynloglikMC(c(1,2,0,1,.25,3.5),Metro2019,10000,3,3,'exp')
 
-
-dynloglikMC = function (x, y, nreps, weight)
+dynloglikMC = function (x, y, nreps, xiInst, betaInst, weight)
 {
   if (weight == 'cau')
   {
@@ -27,7 +28,7 @@ dynloglikMC = function (x, y, nreps, weight)
     f <- function(x) (evir::dgpd(x, xi, 0, beta) - dlnorm(x, mu, sigma)) * atan((x - muc)/tau)
     p <- pcauchy(y, muc, tau)
     temp <- (1 - p) * dlnorm(y, mu, sigma) + p * evir::dgpd(y, xi, 0, beta)
-    Z = nConst_MC(x, nreps,3,3,'cau')
+    Z = nConst_MC(x, nreps, xiInst, betaInst, 'cau')
     llik <- sum(log(temp/Z))
   }
   
@@ -42,7 +43,7 @@ dynloglikMC = function (x, y, nreps, weight)
       exp(-lambda*x)
     p <- pexp(y,lambda)  
     temp <- (1 - p) * dlnorm(y, mu, sigma) + p * evir::dgpd(y, xi, 0, beta)
-    Z = nConst_MC(x, nreps,3,3,'exp')
+    Z = nConst_MC(x, nreps, xiInst, betaInst, 'exp')
     llik <- sum(log(temp/Z))
   }
   return(llik)