+++ NEWS.md
@@ -1,3 +1,9 @@
+# FitDynMix 1.1.0
+
+* In addition to the lognormal-GPD dynamic mixture with Cauchy weight, the package now implements estimation of a lognormal-GPD dynamic mixture with exponential weight.
+
+* Functionalities for noisy cross-entropy estimation have been added.
+
 # FitDynMix 0.1.0
 
 The first release of FitDynMix provides facilities for estimation of a lognormal-GPD dynamic mixture.